DDA-85: Own files for save&read--diograph-json + test refactoring
