+++ electron/channels/save-room-channel.js
@@ -1,4 +1,4 @@
-const { saveDiographJson } = require('../lib/room-util')
+const { saveDiographJson } = require('../lib/save-diograph-json')
 
 exports.saveRoomEventHandler = (event, { path, room: { rootId, diograph } }) =>
   new Promise((resolve, reject) => {
