+++ electron/lib/read-diograph-json.js
@@ -1,5 +1,4 @@
 const fs = require('fs')
-const fsPromise = require('fs').promises
 const backendLogger = require('electron-log')
 
 /**
@@ -41,19 +40,4 @@ exports.readDiographJson = (path) => {
     rootId: undefined,
     diograph: undefined,
   }
-}
-
-/*
- * saveDiographJson
- * @function
- * @param path {string} - Path of the diograph folder
- * @param diograph {object} - Diograph object to be saved to diograph.json
- * @param rootId {object} - RootId to be saved to diograph.json
- * @return {Promise} - Resolves with no arguments upon success
- *
- */
-exports.saveDiographJson = (path, diograph, rootId) => {
-  backendLogger.info('Saving diograph.json to:', path)
-  const fileContent = { rootId, diograph }
-  return fsPromise.writeFile(`${path}/diograph.json`, JSON.stringify(fileContent, null, 2))
 }
